move configuration stuff to config.rs
